+++ NEWS.md
@@ -9,3 +9,4 @@
 * Added `summarise_trait_means` to aid `trait_pivot_wider`
 * Added `summarise_austraits`
 * Added `lookup_trait`
+* Added `print.austraits`